% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{checkInputType}
\alias{checkInputType}
\title{Check audio input type}
\usage{
checkInputType(x)
}
\arguments{
\item{x}{path to a .wav or .mp3 file, Wave object, or a numeric vector
representing the waveform with specified samplingRate}
}
\description{
Internal soundgen function.
}
\details{
Checks the types of audio input to another function, which could be a folder
with audio files, a single file, a Wave object, or a numeric vector. The
purposes of this helper function are to ascertain that there are some valid
inputs and to make a list of valid audio files, if any.
}
\keyword{internal}
