% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{convert_sec_to_hms}
\alias{convert_sec_to_hms}
\title{Print time}
\usage{
convert_sec_to_hms(time_s)
}
\arguments{
\item{time_s}{time (s)}
}
\value{
Returns a character string like "1 h 20 min 3 s"
}
\description{
Internal soundgen function.
}
\details{
Converts time in seconds to time in hh:mm:ss for pretty printing.
}
\examples{
time_start = proc.time()
Sys.sleep(1)
time_diff = as.numeric((proc.time() - time_start)[3])
soundgen:::convert_sec_to_hms(time_diff)
}
\keyword{internal}
