% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\docType{data}
\name{def_form}
\alias{def_form}
\title{Defaults and ranges for formant_app()}
\format{
A matrix with 4 columns:
\describe{
  \item{default}{default value}
  \item{low}{lowest permitted value}
  \item{high}{highest permitted value}
  \item{step}{increment for adjustment}
  ...
}
}
\usage{
def_form
}
\description{
Internal soundgen list of defaults.
}
\details{
A dataset containing defaults and ranges of key variables for formant_app().
Adjust as needed.
}
\keyword{internal}
