% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defaults}
\alias{defaults}
\title{Shiny app defaults}
\format{An object of class \code{list} of length 65.}
\usage{
defaults
}
\description{
A list of default values for Shiny app soundgen_app() - mostly the same as
the defaults for soundgen(). NB: if defaults change, this has to be
updated!!!
}
\keyword{datasets}
