% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\docType{data}
\name{defaults_analyze}
\alias{defaults_analyze}
\title{Defaults and ranges for analyze()}
\format{A matrix with 58 rows and 4 columns:
\describe{
  \item{default}{default value}
  \item{low}{lowest permitted value}
  \item{high}{highest permitted value}
  \item{step}{increment for adjustment}
  ...
}}
\usage{
defaults_analyze
}
\description{
A dataset containing defaults and ranges of key variables for analyze() and
pitch_app(). Adjust as needed.
}
\keyword{datasets}
