% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\docType{data}
\name{defaults_analyze_pitchCand}
\alias{defaults_analyze_pitchCand}
\title{Defaults for plotting with analyze()}
\format{
A dataframe with 8 rows and 5 columns:
\describe{
  \item{method}{pitch tracking method}
  \item{col}{color}
  \item{pch}{point character}
  \item{lwd}{line width}
  \item{lty}{line type}
  ...
}
}
\usage{
defaults_analyze_pitchCand
}
\description{
Default plotting settings for each pitch tracker in analyze() and
pitch_app(). Adjust as needed.
}
\keyword{datasets}
