% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detectNLP_training_nonv}
\alias{detectNLP_training_nonv}
\title{Nonlinear phenomena: Naive Bayes classifier trained on human nonverbal
vocalizations}
\format{
An object of class \code{list} of length 9.
}
\usage{
detectNLP_training_nonv
}
\description{
The results of running \code{\link{naiveBayes_train}} on acoustically
analyzed 969 human nonverbal vocalizations (>83K frames). It is used by
\code{\link{detectNLP}}.
}
\keyword{datasets}
