% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detectNLP_training_synth}
\alias{detectNLP_training_synth}
\title{Nonlinear phenomena: Naive Bayes classifier trained on synthetic sounds}
\format{
An object of class \code{list} of length 9.
}
\usage{
detectNLP_training_synth
}
\description{
The results of running \code{\link{naiveBayes_train}} on 5000 synthetic
sounds with or without NLP created with soundgen(). It is used by
\code{\link{detectNLP}}.
}
\keyword{datasets}
