% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{.getDuration}
\alias{.getDuration}
\title{Get duration per sound}
\usage{
.getDuration(audio, silence = 0.01, rms = list(windowLength = 20, step = 5))
}
\arguments{
\item{audio}{a list returned by \code{readAudio}}

\item{silence}{leading and trailing sections quieter than this proportion of
maximum RMS amplitude are removed when calculating
\code{duration_noSilence} (NULL = don't calculate \code{duration_noSilence}
to save time)}

\item{rms}{a list of control parameters passed to \code{\link{getRMS}}}
}
\description{
Internal soundgen function called by \code{\link{getDuration}}.
}
\keyword{internal}
