% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{.pitchDescriptives}
\alias{.pitchDescriptives}
\title{Pitch descriptives per file}
\usage{
.pitchDescriptives(
  time,
  pitch,
  smoothBW = c(25, 2),
  inflThres = 2,
  plot = FALSE
)
}
\arguments{
\item{time}{a vector of time stamps in s}

\item{pitch}{a vector of pitch values in Hz}

\item{smoothBW}{a vector of bandwidths (Hz) for consecutive smoothing of
input using \code{\link{pitchSmoothPraat}}}

\item{inflThres}{minimum difference (in semitones) between consecutive
extrema to consider them inflections; to apply a different threshold at
each smoothing level, provide \code{inflThres} as a vector of the same
length as \code{smoothBW}}

\item{plot}{if TRUE, plots the inflections for manual verification}
}
\description{
Internal soundgen function.
}
\keyword{internal}
