% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothContours.R
\name{drawContour}
\alias{drawContour}
\title{Draw contour}
\usage{
drawContour(len, anchors, method, valueFloor, valueCeiling, duration_ms = 500)
}
\arguments{
\item{len}{the required length of the output contour. If NULL, it will be
calculated based on the maximum time value (in ms) and \code{samplingRate}}

\item{anchors}{a dataframe of anchors with specified time and amplitude.
\code{achors$time} can be in ms (with len=NULL) or in arbitrary units, eg 0
to 1 (with duration determined by len, which must then be provided in ms).
So anchors$time is assumed to be in ms if len=NULL and relative if len is
specified. \code{anchors$value} can be on any scale.}

\item{method}{smoothing method: 'spline' or 'loess' (default). Consider
'spline' if you have a lot of anchors and not much patience, since it is
much faster than 'loess', but it may produce weird results when there are
only a few anchors}

\item{valueFloor}{lower/upper bounds for the contour}

\item{valueCeiling}{lower/upper bounds for the contour}
}
\description{
Internal soundgen function
}
\details{
The core part of getSmoothContour() that actually performs the interpolation
between anchors.
}
