% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{drawFreqAxis}
\alias{drawFreqAxis}
\title{Draw frequency axis}
\usage{
drawFreqAxis(y, ylim = range(y), yScale, nLbls = 5, y_Hz = TRUE, ...)
}
\arguments{
\item{y}{frequency values}

\item{ylim}{range of frequency values}

\item{yScale}{scale of frequency representation}

\item{nLbls}{number of frequency labels}

\item{y_Hz}{show frequency in Hz (TRUE) or kHz (FALSE)}

\item{...}{other graphical parameters passed to axis()}
}
\description{
Internal soundgen function
}
\details{
Helper function for adding a customized frequency axis to the spectrogram
}
\keyword{internal}
