% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{fart}
\alias{fart}
\title{Fart}
\usage{
fart(glottis = c(350, 700), glottisAnchors = "deprecated", pitch = 75,
  pitchAnchors = "deprecated", temperature = 0.25, sylLen = 600,
  rolloff = -20, samplingRate = 16000, play = FALSE, plot = FALSE)
}
\arguments{
\item{glottis}{anchors for specifying the proportion of a
glottal cycle with closed glottis, \% (0 = no modification, 100 = closed
phase as long as open phase); numeric vector or dataframe specifying time
and value (anchor format)}

\item{glottisAnchors}{anchors for specifying the proportion of a
glottal cycle with closed glottis, \% (0 = no modification, 100 = closed
phase as long as open phase); numeric vector or dataframe specifying time
and value (anchor format)}

\item{pitch}{a numeric vector of f0 values in Hz or a dataframe
specifying the time (ms or 0 to 1) and value (Hz) of each anchor, hereafter
"anchor format". These anchors are used to create a smooth contour of
fundamental frequency f0 (pitch) within one syllable}

\item{pitchAnchors}{a numeric vector of f0 values in Hz or a dataframe
specifying the time (ms or 0 to 1) and value (Hz) of each anchor, hereafter
"anchor format". These anchors are used to create a smooth contour of
fundamental frequency f0 (pitch) within one syllable}

\item{temperature}{hyperparameter for regulating the amount of stochasticity
in sound generation}

\item{sylLen}{syllable length, ms (not vectorized)}

\item{rolloff}{rolloff of harmonics in source spectrum, dB/octave (not
vectorized)}

\item{samplingRate}{sampling frequency, Hz}

\item{play}{if TRUE, plays the synthesized sound. In case of errors, try
setting another default player for \code{\link[tuneR]{play}}}

\item{plot}{if TRUE, plots the waveform}
}
\value{
Returns a normalized waveform.
}
\description{
While the same sounds can be created with soundgen(), this dedicated
facetious function produces the same effect more efficiently and with very
few control parameters. With default settings, execution time is ~ 10 ms per
second of audio sampled at 16000 Hz. Principle: creates separate glottal
cycles with harmonics, but no formants. See \code{\link{soundgen}} for more
details.
}
\examples{
f = fart()
# playme(f)
}
