% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{findElbow}
\alias{findElbow}
\title{Find the elbow of a screeplot or similar}
\usage{
findElbow(d, plot = FALSE)
}
\arguments{
\item{d}{dataframe containing x and y coordinates of the points}
}
\description{
Adapted from
https://stackoverflow.com/questions/2018178/finding-the-best-trade-off-point-on-a-curve
Algorithm: draw a straight line between the two endpoints and find the point
furthest from this line.
}
\examples{
y = c(10, 11, 8, 4, 2, 1.5, 1, 0.7, .5, .4, .3)
soundgen:::findElbow(data.frame(x = seq_along(y), y = y), plot = TRUE)
}
\keyword{internal}
