% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_segment.R
\name{findSyllables}
\alias{findSyllables}
\title{Find syllables}
\usage{
findSyllables(ampl, step, windowLength, threshold, shortestSyl, shortestPause)
}
\arguments{
\item{step}{time difference between two points in the envelope (ms)}

\item{windowLength}{length of FFT window, ms}

\item{threshold}{all continuous segments above this value are considered to
be syllables}

\item{shortestSyl}{minimum acceptable length of syllables, ms}

\item{shortestPause}{minimum acceptable break between syllables, ms
(syllables separated by shorter pauses are merged)}

\item{envelope}{downsampled amplitude envelope: dataframe with columns "time"
and "value"}
}
\value{
Returns a dataframe with timing of syllables.
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{segment}}.
}
\keyword{internal}
