% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude.R
\name{flatEnvSound}
\alias{flatEnvSound}
\title{Flat envelope per sound}
\usage{
flatEnvSound(
  audio,
  compression = 1,
  method = c("hil", "rms", "peak")[1],
  windowLength = 50,
  windowLength_points = NULL,
  killDC = FALSE,
  dynamicRange = 40,
  plot = FALSE,
  saveAudio = NULL
)
}
\arguments{
\item{audio}{a list returned by \code{readAudio}}

\item{compression}{the amount of compression to apply: 0 = none, 1 = maximum}

\item{method}{hil = Hilbert envelope, rms = root mean square amplitude, peak
= peak amplitude per window}

\item{windowLength}{the length of smoothing window, ms}

\item{windowLength_points}{the length of smoothing window, points. If
specified, overrides \code{windowLength}}

\item{killDC}{if TRUE, dynamically removes DC offset or similar deviations of
average waveform from zero (see examples)}

\item{dynamicRange}{parts of sound quieter than \code{-dynamicRange} dB will
not be amplified}

\item{plot}{if TRUE, plots the original sound, the smoothed envelope, and
the compressed sound}

\item{saveAudio}{full path to the folder in which to save the compressed
sound(s)}
}
\description{
Internal soundgen function
}
\keyword{internal}
