% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formants.R
\name{formantDev}
\alias{formantDev}
\title{Formant deviation from schwa}
\usage{
formantDev(formants, vocalTract = NULL, speedSound = 35400)
}
\arguments{
\item{formants}{a numeric vector of observed formant frequencies, Hz}

\item{vocalTract}{the length of vocal tract, cm}

\item{speedSound}{speed of sound in warm air, cm/s. Stevens (2000) "Acoustic
phonetics", p. 138}
}
\value{
Returns a vector of percentages indicating how much a given formant
  is higher (positive values) or lower (negative values) than expected based
  on the estimated formant dispersion.
}
\description{
Calculates the deviation of formant frequencies from the neutral values
expected in a perfectly cylindrical vocal tract (schwa sound). Algorithm:
estimates formant dispersion based on all observed formant frequencies,
calculates theoretically expected formant frequencies for a schwa sound in a
vocal tract with this formant dispersion, and compares them with the observed
frequencies. Basically, we estimate vocal tract length and see if each
formant is higher or lower than expected for this vocal tract. For this to
work, we have to know either the frequencies of enough formants (not just the
first two) or the true length of the vocal tract.
}
\examples{
# Formant frequencies in three vocalizations by the same male speaker:
formants_a = c(860, 1430, 2900, 4200, 5200)
estimateVTL(formants_a)  # 15.2 cm
formantDev(formants_a)
# [a]: F1 much higher than expected, F2 slightly lower

formants_i = c(300, 2700, 3400, 4400, 5300, 6400)
estimateVTL(formants_i)  # 14.5 cm
formantDev(formants_i)
# [i]: very low F1, very high F2

formants_roar = c(550, 1000, 1460, 2280, 3350, 4300, 4900, 5800, 6900, 7900)
estimateVTL(formants_roar)  # 22.5 cm (!)
# (lowered larynx and rounded lips to exaggerate apparent size)
formantDev(formants_roar)
# roar: high F1 and low F2-F4
estimateVTL(formants_roar[1:4]) # based on F1-F4, VTL is almost 28 cm!
# Since the lowest formants are the most salient,
the apparent size is exaggerated even further
}
