% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{formatPitchManual}
\alias{formatPitchManual}
\title{Format pitchManual}
\usage{
formatPitchManual(pitchManual)
}
\arguments{
\item{pitchManual}{dataframe produced by analyze() or pitch_app(), path to a
.csv file in which this dataframe is stored, a named list with a numeric
vector of pitch values per sound, or a numeric vector}
}
\value{
A named list of pitch contours.
}
\description{
Internal soundgen function
}
\examples{
soundgen:::formatPitchManual(c(NA, 120, 180, NA))
soundgen:::formatPitchManual('NA, 120, 180, NA')
soundgen:::formatPitchManual(list('myfile.wav' = c(NA, 120, 180, NA)))
soundgen:::formatPitchManual(data.frame(file = c('file1.wav', 'file2.wav'),
                                        pitch = c('NA, 120', '180, NA')))
soundgen:::formatPitchManual('adja')
}
\keyword{internal}
