% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{ftwindow_modif}
\alias{ftwindow_modif}
\title{Fourier transform windows (seewave)}
\usage{
ftwindow_modif(wl, wn = "gaussian")
}
\arguments{
\item{wl}{window length, in points}

\item{wn}{window type (defaults to gaussian)}
}
\description{
Internal soundgen function
}
\details{
Generates different Fourier Transform windows. Just like
\code{\link[seewave]{ftwindow}}, but with the addition of a gaussian window.
}
\keyword{internal}
