% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{gaussian.w}
\alias{gaussian.w}
\title{Gaussian window}
\usage{
gaussian.w(n)
}
\arguments{
\item{n}{window length, in points}
}
\description{
Internal soundgen function.
}
\details{
Generates a gaussian window of length n. Based on the formula by P. Boersma (PRAAT)
}
\keyword{internal}
