% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{generatePath}
\alias{generatePath}
\title{Generate path}
\usage{
generatePath(path, pitchCands, ...)
}
\arguments{
\item{path}{currently evaluated path}

\item{pitchCands}{a matrix of multiple pitch candidates per fft frame. Each
column is one fft frame, each row is one candidate (the last row is always
"manual")}

\item{...}{nothing really, but otherwise optim() complains}
}
\description{
Internal soundgen function.
}
\details{
Internal helper function for postprocessing of pitch contours called by
\code{\link{pathfinding_slow}}. Generates proposals for new paths through
pitch candidates. It gives up and returns NA after 100 attempts, which stops
annealing - so the adaptation of pitch contour doesn't happen
}
\keyword{internal}
