% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchTrackers.R
\name{getCPP}
\alias{getCPP}
\title{Get Cepstral Peak Prominence}
\usage{
getCPP(frame, samplingRate, pitch, bin, prox_semitones = 3)
}
\arguments{
\item{frame}{the abs spectrum of a frame, as returned by
\code{\link[stats]{fft}}}

\item{samplingRate}{sampling rate (Hz)}

\item{pitch}{pitch of this frame, Hz}

\item{bin}{spectral frequency bin width, Hz}
}
\value{
Returns either NA or CPP in dB.
}
\description{
Internal soundgen function.
}
\details{
Calculates Cepstral Peak Prominence from the spectrum of an STFT frame.
}
\keyword{internal}
