% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudness.R
\name{getLoudnessFolder}
\alias{getLoudnessFolder}
\title{Get loudness folder}
\usage{
getLoudnessFolder(...)
}
\arguments{
\item{...}{any input parameters}
}
\description{
Deprecated; use \code{\link{getLoudness}} instead
}
