% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude.R
\name{getRMSSound}
\alias{getRMSSound}
\title{RMS amplitude per sound}
\usage{
getRMSSound(
  sound,
  samplingRate,
  scale = NULL,
  windowLength = 50,
  step = NULL,
  overlap = 75,
  killDC = FALSE,
  normalize = TRUE,
  windowDC = 200,
  plot = TRUE,
  xlab = "",
  ylab = "",
  type = "b",
  col = "green",
  lwd = 2,
  ...
)
}
\arguments{
\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than an audio file or Wave object)}

\item{scale}{maximum possible amplitude of input used for normalization (not
needed for audio files)}

\item{windowLength}{length of FFT window, ms}

\item{step}{you can override \code{overlap} by specifying FFT step, ms (NB:
because digital audio is sampled at discrete time intervals of
1/samplingRate, the actual step and thus the time stamps of STFT frames
will be slightly different, eg 24.98866 instead of 25.0 ms)}

\item{overlap}{overlap between successive FFT frames, \%}

\item{killDC}{if TRUE, removed DC offset (see also \code{\link{flatEnv}})}

\item{normalize}{if TRUE, RMS amplitude is normalized to [0, 1]}

\item{windowDC}{the window for calculating DC offset, ms}

\item{plot}{if TRUE, plot a contour of RMS amplitude}

\item{xlab}{general graphical parameters}

\item{ylab}{general graphical parameters}

\item{type}{graphical parameters pertaining to the RMS envelope}

\item{col}{graphical parameters pertaining to the RMS envelope}

\item{lwd}{graphical parameters pertaining to the RMS envelope}

\item{...}{other graphical parameters}
}
\description{
Internal soundgen function called by \code{\link{getRMS}}.
}
\seealso{
\code{\link{getRMSFolder}}
  \code{\link{analyze}}\code{\link{getLoudness}}
}
\keyword{internal}
