% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{getSmoothSpectrum}
\alias{getSmoothSpectrum}
\title{Get smooth spectrum}
\usage{
getSmoothSpectrum(
  sound,
  samplingRate,
  len,
  loessSpan,
  plot = FALSE,
  xlab = "Frequency, kHz",
  ylab = "dB",
  type = "l",
  ...
)
}
\arguments{
\item{sound}{the audio (numeric, any scale)}

\item{samplingRate}{the sampling rate, Hz}

\item{len}{the desired resolution of the output}

\item{loessSpan}{passed to loess to control the amount of smoothing (.01 =
minimal smoothing, 1 = strong smoothing)}
}
\description{
Internal soundgen function.
}
\examples{
s = soundgen(sylLen = 100, pitch = 500, addSilence = FALSE)
soundgen:::getSmoothSpectrum(s, 16000, len = 500, loessSpan = .01, plot = TRUE)
soundgen:::getSmoothSpectrum(s, 16000, len = 500, loessSpan = .1, plot = TRUE)
soundgen:::getSmoothSpectrum(s, 16000, len = 500, loessSpan = .5, plot = TRUE)
soundgen:::getSmoothSpectrum(s, 16000, len = 500, loessSpan = 1, plot = TRUE)
}
\keyword{internal}
