% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{getSpectralFlux}
\alias{getSpectralFlux}
\title{Get spectral flux}
\usage{
getSpectralFlux(s)
}
\arguments{
\item{s}{raw spectrogram (not normalized): rows = frequency bins, columns = STFT frames}
}
\value{
vector of length ncol(s)
}
\description{
Internal soundgen function
}
\details{
Calculates spectral flux: the average change across all spectral bins from
one STFT frame to the next. Spectra are normalized in each frame, so
amplitude changes have no effect on flux.
}
\keyword{internal}
