% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisal.R
\name{getSurprisal_matrix}
\alias{getSurprisal_matrix}
\title{Get surprisal per matrix}
\usage{
getSurprisal_matrix(x, win)
}
\arguments{
\item{x}{input matrix such as a spectrogram (columns = time, rows =
frequency)}

\item{win}{length of analysis window}
}
\description{
Internal soundgen function called by \code{\link{getSurprisal}}.
}
\keyword{internal}
