% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralDescr.R
\name{harmHeight_dif}
\alias{harmHeight_dif}
\title{Height of harmonics: difference method}
\usage{
harmHeight_dif(
  frame_dB,
  pitch,
  bin,
  freqs,
  harmThres = 3,
  harmTol = 0.25,
  harmPerSel = 5,
  plot = FALSE
)
}
\arguments{
\item{pitch}{the final pitch estimate for the current frame}

\item{bin}{spectrogram bin width, Hz}

\item{freqs}{frequency per bin of spectrogram}

\item{harmThres}{minimum height of spectral peak, dB}

\item{harmTol}{maximum tolerated deviation of peak frequency from multiples
of f0, proportion of f0}

\item{harmPerSel}{the number of harmonics per sliding selection}

\item{plot}{if TRUE, produces a plot of spectral peaks}
}
\description{
Internal soundgen function
}
\details{
Estimates how far harmonics reach in the spectrum by analyzing the typical
distances between spectral peaks in different frequency regions.
}
\keyword{internal}
