% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralDescr.R
\name{harmHeight_peaks}
\alias{harmHeight_peaks}
\title{Height of harmonics: peaks method}
\usage{
harmHeight_peaks(
  frame_dB,
  pitch,
  bin,
  freqs,
  harmThres = 3,
  harmTol = 0.25,
  plot = FALSE
)
}
\arguments{
\item{pitch}{the final pitch estimate for the current frame}

\item{bin}{spectrogram bin width, Hz}

\item{freqs}{frequency per bin of spectrogram}

\item{harmThres}{minimum height of spectral peak, dB}

\item{harmTol}{maximum tolerated deviation of peak frequency from multiples
of f0, proportion of f0}

\item{plot}{if TRUE, produces a plot of spectral peaks}
}
\description{
Internal soundgen function
}
\details{
Estimates how far harmonics reach in the spectrum by checking how many
spectral peaks we can find close to multiples of f0.
}
\keyword{internal}
