% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{htmlPlots}
\alias{htmlPlots}
\title{HTML for clickable plots}
\usage{
htmlPlots(
  x,
  savePlots,
  changesAudio,
  suffix,
  extension = "png",
  width = "900px"
)
}
\arguments{
\item{width}{default flex-basis, ie approximate default width of each image}

\item{myfolder}{full path to target folder, without a '/' at the end}

\item{myfiles}{a list of full names of files (with paths and extensions)}
}
\description{
Internal soundgen function
}
\details{
Writes an html file for displaying clickable plots in a browser.
}
\examples{
\dontrun{
## Functions that modify audio
# relative paths
flatEnv('/home/allgoodguys/Downloads/temp', savePlots = '')
flatEnv('/home/allgoodguys/Downloads/temp',
  savePlots = '/home/allgoodguys/Downloads/temp/compr',
  saveAudio = '/home/allgoodguys/Downloads/temp/compr')

# absolute paths
flatEnv('/home/allgoodguys/Downloads/temp', savePlots = '',
  saveAudio = '/home/allgoodguys/Downloads/temp/compr')
flatEnv('/home/allgoodguys/Downloads/temp',
  savePlots = '~/Downloads/temp/plots_compr')


## Functions that only analyze audio
getRMS('/home/allgoodguys/Downloads/temp', savePlots = '')
getRMS('/home/allgoodguys/Downloads/temp',
  savePlots = '~/Downloads/temp/plots_rms')
}
}
\keyword{internal}
