% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_scales.R
\name{hz2mel}
\alias{hz2mel}
\title{Convert Hz to mel}
\usage{
hz2mel(f, htk = TRUE)
}
\arguments{
\item{f}{frequency, Hz}

\item{htk}{if TRUE, uses an alternative formula 2595 * log10(1 + f/700)}
}
\description{
Internal soundgen function: a temporary fix needed because tuneR::hz2mel
doesn't accept NAs or vectors.
}
\examples{
soundgen:::hz2mel(c(440, 220, NA))
freq = 1:10000
plot(freq, soundgen:::hz2mel(freq), type = 'l')
}
\keyword{internal}
