% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertModulSpec.R
\name{invertModulationSpectrum}
\alias{invertModulationSpectrum}
\title{Invert modulation spectrum}
\usage{
invertModulationSpectrum(ms)
}
\description{
samplingRate = 16000
s = soundgen(sylLen = 1500, pitch = 200, amFreq = 25, amDep = 50, samplingRate = samplingRate)
playme(s, samplingRate)
spec = spectrogram(s, samplingRate, windowLength = 25, overlap = 80, wn = 'hanning', osc = TRUE, padWithSilence = FALSE)
s_rev = invertSpectrogram(spec, samplingRate = samplingRate, windowLength = 25, overlap = 80, wn = 'hanning')
  playme(s_rev, samplingRate)
  spectrogram(s_rev, samplingRate, osc = TRUE)
}
\details{
ms = modulationSpectrum(s, samplingRate = samplingRate, windowLength = 25, overlap = 80, wn = 'hanning', maxDur = Inf, logSpec = FALSE, power = NA, logWarp = NA, returnComplex = TRUE, kernelSize = 1)$complex
ms = specToMS(spec, samplingRate = samplingRate)
image(x = as.numeric(colnames(ms)), y = as.numeric(rownames(ms)), z = t(log(abs(ms))))

# Filter as needed - for ex., remove AM > 3 Hz
am = as.numeric(colnames(ms))
fm = as.numeric(rownames(ms))
idx_row = 1:nrow(ms)
idx_col = which(abs(am) > 5)
ms_filt = ms
ms_filt[idx_row, idx_col] = 0
image(t(log(abs(ms_filt))))

# Convert back to a spectrogram
spec_filt = msToSpec(ms_filt, samplingRate = samplingRate)
image(t(log(abs(spec_filt))))

Invert the spectrogram
s_filt = invertSpectrogram(abs(spec_filt), samplingRate = samplingRate, windowLength = 25, overlap = 80, initialPhase = 'spsi', nIter = 50)

# Compare with the original
playme(s, samplingRate)
spectrogram(s, samplingRate, osc = TRUE)
playme(s_filt, samplingRate)
spectrogram(s_filt, samplingRate, osc = TRUE)
}
