% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipa}
\alias{ipa}
\title{Manually measured frequencies of the first four formants in 21 vowels from
"Seeing Speech" (http://www.seeingspeech.ac.uk)}
\format{
An object of class \code{data.frame} with 21 rows and 9 columns.
}
\usage{
ipa
}
\description{
A dataframe of 21 rows (one per vowel) and 9 columns: "ipa" = vowel, "F1...4"
= measured formant frequencies (Hz) in the initial stable region of each
vowel, "F1Rel...F4Rel" = formant frequencies relative to the neutral
equidistant formant frequencies in the schwa (semitones)
}
\examples{
plot(ipa$F1, ipa$F2, type = 'n')
text(ipa$F1, ipa$F2, label = ipa$ipa)
}
\keyword{datasets}
