% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{isCentral.localMax}
\alias{isCentral.localMax}
\title{Simple peak detection}
\usage{
isCentral.localMax(x, threshold)
}
\arguments{
\item{x}{input vector}

\item{threshold}{threshold for peak detection}
}
\description{
Internal soundgen function.
}
\examples{
soundgen:::isCentral.localMax(c(1,1,3,2,1), 2.5)
}
\keyword{internal}
