% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinPred.R
\name{isNeighbour_mod}
\alias{isNeighbour_mod}
\title{Is neighbor modified}
\usage{
isNeighbour_mod(v1, v2, embedding.dim, radius)
}
\arguments{
\item{v1, v2, embedding.dim, radius}{see nonlinearTseries:::isNeighbour}
}
\description{
Just a copy of nonlinearTseries:::isNeighbour to avoid a note in
CMD check when calling an unexported function from another package
}
\keyword{internal}
