% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftPitch.R
\name{istft_mod}
\alias{istft_mod}
\title{Modified istft}
\usage{
istft_mod(
  stft,
  f,
  wl,
  ovlp = 75,
  wn = "hanning",
  mult_short = 1,
  mult_long = 1
)
}
\arguments{
\item{stft, f, wl, ovlp, wn}{see seewave:::istft()}

\item{mult_short}{stretch factor of length \code{ncol(stft) - 1}}

\item{mult_long}{stretch factor of length \code{ncol(stft)}}
}
\description{
Internal soundgen function. Similar to seewave:::istft(), but adapted to work
with time-variable step sizes in the context of dynamic pitch shifting. Only
call it for dynamic istft because the original seewave function should be a
bit faster for static.
}
\keyword{internal}
