% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{jet.col}
\alias{jet.col}
\title{Matlab colors}
\usage{
jet.col(n = 100, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transparency}
}
\description{
Internal soundgen function for generating a Matlab-like palette
(=plot3D::jet.col).
}
\keyword{internal}
