% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{listDepth}
\alias{listDepth}
\title{List depth}
\usage{
listDepth(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Internal soundgen function
}
\details{
Returns the depth of list structure. See https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r
}
\keyword{internal}
