% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulationSpectrum.R
\name{logWarpMS}
\alias{logWarpMS}
\title{Log-warp a modulation spectrum}
\usage{
logWarpMS(x, logWarp)
}
\arguments{
\item{x}{a modulation spectrum: rows = FM, cols = AM}
}
\description{
Internal soundgen function
}
\details{
Log-warps a modulation spectrum along time dimension
}
\examples{
a = matrix(1:44, ncol = 11)
colnames(a) = -5:5
soundgen:::logWarpMS(a, logWarp = 2)
}
\keyword{internal}
