% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{matchColumns}
\alias{matchColumns}
\title{Match number of columns}
\usage{
matchColumns(matrix_short, nCol, padWith = 0)
}
\arguments{
\item{matrix_short}{input matrix}

\item{nCol}{the required number of columns}

\item{padWith}{the value to pad with, normally \code{0} or \code{NA}}
}
\description{
Internal soundgen function
}
\details{
Adds columns of zeros or NA to a matrix (attaching them both left and
right), so that the new number of columns = \code{len}
}
\examples{
a = matrix(1:9, nrow = 3)
soundgen:::matchColumns(a, nCol = 6, padWith = NA)
}
\keyword{internal}
