% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_segment.R
\name{mergeSyllables}
\alias{mergeSyllables}
\title{Merge syllables}
\usage{
mergeSyllables(syllables, shortestPause)
}
\arguments{
\item{syllables}{a dataframe listing syllables with start and end}

\item{shortestPause}{minimum acceptable break between syllables, ms.
Syllables separated by less time are merged. To avoid merging, specify
\code{shortestPause = NA}}
}
\description{
Internal soundgen function.
}
\details{
Merges syllables if they are separated by less than \code{shortestPause ms}. Called by \code{\link{findSyllables}}.
}
\keyword{internal}
