% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{Mode}
\alias{Mode}
\title{Modified mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Internal soundgen function
}
\details{
Internal helper function for spectral (~BaNa) pitch tracker. NOT quite the same as simply mode(x).
}
\examples{
soundgen:::Mode(c(1, 2, 3))  # if every element is unique, return the smallest
soundgen:::Mode(c(1, 2, 2, 3))
}
\keyword{internal}
