% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulationSpectrum.R
\name{modulationSpectrumFolder}
\alias{modulationSpectrumFolder}
\title{Modulation spectrum folder}
\usage{
modulationSpectrumFolder(...)
}
\arguments{
\item{...}{any input parameters}
}
\description{
Deprecated; use \code{\link{modulationSpectrum}} instead
}
