% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{na.trim}
\alias{na.trim}
\title{Trim leading and trailing NAs}
\usage{
na.trim(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Internal soundgen function. Nearly 10 times faster than zoo::na.trim.
Slightly slower solution: a[!cumprod(is.na(a)) &
rev(!cumprod(is.na(rev(a))))]. See
https://stackoverflow.com/questions/42759027/remove-leading-and-trailing-na
}
\examples{
soundgen:::na.trim(c(NA, NA, 1:10, NA, 21:25, NA))
}
\keyword{internal}
