% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{naiveBayes_likelihood}
\alias{naiveBayes_likelihood}
\title{Naive Bayes likelihood}
\usage{
naiveBayes_likelihood(
  d,
  nObs = nrow(d),
  mod_train,
  class_names,
  nClasses = length(class_names),
  like_names,
  predictors,
  nPredictors = length(predictors)
)
}
\arguments{
\item{d}{dataframe containing the observations}

\item{nObs}{the number of observations}

\item{mod_train}{the output of naiveBayes_train()}

\item{class_names}{names of outcome classes}

\item{nClasses}{the number of outcome classes}

\item{like_names}{the names of variables holding likelihoods}

\item{predictors}{the names of predictor variables}

\item{nPredictors}{the number of predicto variables}
}
\description{
Internal soundgen function
}
\details{
A Helper function called by \code{\link{naiveBayes}} to calculate the
likelihood of each observation. Algorithm: for each predictor and class, the
likelihood is dnorm(observation, mean_per_class, sd_per_class). I tried
non-Gaussian probability distributions (Student's t to accommodate outliers),
but Gaussian actually seems to be more robust.
}
\keyword{internal}
