% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{osc_dB}
\alias{osc_dB}
\title{Oscillogram dB}
\usage{
osc_dB(x, range = 60, maxAmpl = NULL, samplingRate = NULL, plot = TRUE,
  xlab = NULL, ylab = "dB", bty = "n", midline = TRUE, ...)
}
\arguments{
\item{x}{path to a .wav file or a vector of amplitudes with specified
samplingRate}

\item{range}{dynamic range of the oscillogram, dB}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than a .wav file)}

\item{plot}{if TRUE, plots the oscillogram; if FALSE, simply returns the
log-transformed waveform}

\item{xlab, ylab}{axis labels}

\item{bty}{box type (see `?par`)}

\item{midline}{if TRUE, draws a line at 0 dB}

\item{...}{Other graphical parameters passed on to `plot()`}
}
\value{
Returns the input waveform on a dB scale: a centered vector with
  range from `-range` to `range`.
}
\description{
Plots the oscillogram (waveform) of a sound on a logarithmic scale, in dB.
Analogous to "Waveform (dB)" view in Audacity.
}
\details{
Algorithm: centers and normalizes the sound, then takes a logarithm of the
positive part and a flipped negative part.
}
\examples{
sound = sin(1:2000/10) *
        getSmoothContour(anchors = c(1, .01, .5), len = 2000)

# Oscillogram on a linear scale
plot(sound, type = 'l')
# or, for fancy plotting options: seewave::oscillo(sound, f = 1000)

# Oscillogram on a dB scale
o = osc_dB(sound)

# Time in ms if samplingRate is specified
o = osc_dB(sound, samplingRate = 5000)

# Assuming that the waveform can range up to 50 instead of 1
o = osc_dB(sound, maxAmpl = 50)

# Embellish and customize the plot
o = osc_dB(sound, samplingRate = 1000, midline = FALSE,
           main = 'My waveform', col = 'blue')
abline(0, 0, col = 'orange', lty = 3)
}
