% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\docType{data}
\name{permittedValues}
\alias{permittedValues}
\title{Defaults and ranges for soundgen()}
\format{
A matrix with 58 rows and 4 columns:
\describe{
  \item{default}{default value}
  \item{low}{lowest permitted value}
  \item{high}{highest permitted value}
  \item{step}{increment for adjustment}
  ...
}
}
\usage{
permittedValues
}
\description{
A dataset containing defaults and ranges of key variables for soundgen() and
soundgen_app().
Adjust as needed.
}
\keyword{datasets}
