% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftPitch.R
\name{phasePropagate}
\alias{phasePropagate}
\title{Propagate phase}
\usage{
phasePropagate(i, dp_hor, dp_ver, magn, phase_new, tol, bin_width, alpha)
}
\arguments{
\item{i}{analyzed frame}

\item{dp_hor, dp_ver}{time and frequency partical derivatives of phase}

\item{magn}{phase and magnitude of a spectrogram}

\item{phase_new}{matrix for storing the new phase}

\item{tol}{tolerance of "vocoder done right" algorithm}

\item{bin_width}{width of frequency bin, Hz}

\item{alpha}{stretch factor}
}
\description{
Internal soundgen function called by dPhase(). Propagates phase using the
"vocoder done right" algorithm, as in Prusa & Holighaus 2017 "Phase vocoder
done right".
}
\keyword{internal}
