% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phonCurves}
\alias{phonCurves}
\title{Equal loudness curves for converting dB SPL to phon}
\format{An object of class \code{list} of length 41.}
\usage{
phonCurves
}
\description{
A list of dataframes, from 0 to 90 phon (see the name of each dataframe).
Translated from the matlab implementation by Jeff Tackett (03/01/05)
available from https://www.mathworks.com/matlabcentral/fileexchange/
7028-iso-226-equal-loudness-level-contour-signal
}
\keyword{datasets}
