% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitchContour}
\alias{pitchContour}
\title{Manually corrected pitch contours in 260 sounds}
\format{An object of class \code{data.frame} with 260 rows and 2 columns.}
\usage{
pitchContour
}
\description{
A dataframe of 260 rows and two columns: "file" for filename in the corpus
(Anikin & Persson, 2017) and "pitch" for pitch values per frame. The corpus
can be downloaded from http://cogsci.se/publications.html
}
\keyword{datasets}
