% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitchManual}
\alias{pitchManual}
\title{Manual pitch estimation in 260 sounds}
\format{An object of class \code{numeric} of length 260.}
\usage{
pitchManual
}
\description{
A vector of manually verified pitch values per sound in the corpus of 590 human non-linguistic emotional vocalizations from Anikin & Persson (2017). The corpus can be downloaded from http://cogsci.se/personal/results/01_anikin-persson_2016_naturalistics-non-linguistic-vocalizations/01_anikin-persson_2016_naturalistic-non-linguistic-vocalizations.html
}
\keyword{datasets}
