% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{princarg}
\alias{princarg}
\title{Principal argument}
\usage{
princarg(x)
}
\arguments{
\item{x}{real number representing phase angle}
}
\description{
Internal soundgen function.
}
\details{
Recalculates the phase of complex numbers to be within the interval from -pi to pi.
}
\keyword{internal}
