% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pseudoLog}
\alias{pseudoLog}
\title{Pseudolog}
\usage{
pseudoLog(x, sigma, base)
}
\arguments{
\item{x}{numeric vector to transform}

\item{sigma}{scaling factor for the linear part}

\item{approximate}{logarithm base used}
}
\description{
Internal soundgen function
}
\details{
From scales::pseudo_log_trans()
}
\examples{
a = -30:30
plot(a, soundgen:::pseudoLog(a, sigma = 1, base = 2))
plot(a, soundgen:::pseudoLog(a, sigma = 5, base = 2))
plot(a, soundgen:::pseudoLog(a, sigma = .1, base = 2))
}
\seealso{
\code{\link{pseudoLog_undo}}
}
\keyword{internal}
