% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{readAudio}
\alias{readAudio}
\title{Read audio}
\usage{
readAudio(
  x,
  input = checkInputType(x),
  i,
  samplingRate = NULL,
  scale = NULL,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{x}{audio input (only used for Wave objects or numeric vectors)}

\item{input}{a list returned by \code{\link{checkInputType}}}

\item{i}{iteration}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than an audio file or Wave object)}

\item{scale}{maximum possible amplitude of input used for normalization of
input vector (only needed if \code{x} is a numeric vector, rather than an
audio file or Wave object)}

\item{from, to}{if NULL (default), analyzes the whole sound, otherwise
from...to (s)}
}
\description{
Internal soundgen function.
}
\keyword{internal}
