% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{sampleModif}
\alias{sampleModif}
\title{sampleModif}
\usage{
sampleModif(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{other arguments passed to \code{sample}}
}
\description{
Internal soundgen function
}
\details{
Same as \code{\link[base]{sample}}, but without defaulting to x = 1:x if
length(x) = 1. See
https://stackoverflow.com/questions/7547758/using-sample-with-sample-space-size-1
}
\examples{
soundgen:::sampleModif(x = 3, n = 1)
# never returns 1 or 2: cf. sample(x = 3, n = 1)
}
\keyword{internal}
