% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segmentManual}
\alias{segmentManual}
\title{Manual counts of syllables in 260 sounds}
\format{
An object of class \code{numeric} of length 260.
}
\usage{
segmentManual
}
\description{
A vector of the number of syllables in the corpus of 260 human non-linguistic
emotional vocalizations from Anikin & Persson (2017). The corpus can be
downloaded from http://cogsci.se/publications.html
}
\keyword{datasets}
