% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM.R
\name{selfsim}
\alias{selfsim}
\title{Compute self-similarity}
\usage{
selfsim(m, norm = FALSE, simil = c("cosine", "cor")[1], win = 1)
}
\arguments{
\item{m}{input matrix such as a spectrogram}

\item{norm}{if TRUE, the spectrum of each STFT frame is normalized}

\item{simil}{method for comparing frames: "cosine" = cosine similarity, "cor"
= Pearson's correlation}

\item{win}{the length of window for averaging self-similarity, frames}
}
\value{
Returns a square self-similarity matrix.
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{ssm}}.
}
\keyword{internal}
