% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertModulSpec.R
\name{specToMS}
\alias{specToMS}
\title{Spectrogram to modulation spectrum}
\usage{
specToMS(spec, samplingRate = NULL, step = NULL)
}
\arguments{
\item{spec}{target spectrogram (numeric matrix, frequency in rows, time in
columns)}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than an audio file)}

\item{step}{you can override \code{overlap} by specifying FFT step, ms}
}
\value{
Returns a MS - matrix of complex values of the same dimension as
  spec, with AM in rows and FM in columns.
}
\description{
Internal soundgen function.
}
\details{
Takes a spectrogram (either complex or magnitude) and returns a MS with
proper row and column labels.
}
\keyword{internal}
