% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spreadSpecCoef}
\alias{spreadSpecCoef}
\title{Matrix of scaling coefficients for spreading bark spectrum}
\format{An object of class \code{matrix} with 27 rows and 27 columns.}
\usage{
spreadSpecCoef
}
\description{
See Wonho (1999) "Enhanced modified bark spectral distortion (EMBSD)".
Original: Schroeder, M. R., Atal, B. S., & Hall, J. L. (1979). Optimizing
digital speech coders by exploiting masking properties of the human ear. The
Journal of the Acoustical Society of America, 66(6), 1647-1652. Called by
\code{\link{spreadSpec}}
}
\keyword{datasets}
