% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{summarizeAnalyze}
\alias{summarizeAnalyze}
\title{Summarize the output of analyze()}
\usage{
summarizeAnalyze(
  result,
  summaryFun = c("mean", "sd"),
  var_noSummary = c("duration", "duration_noSilence", "voiced", "time")
)
}
\arguments{
\item{result}{dataframe returned by analyze(summary = FALSE)}

\item{summaryFun}{summary functions}

\item{var_noSummary}{variables that should not be summarized}
}
\description{
Internal soundgen function
}
\keyword{internal}
