% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{updateAnalyze}
\alias{updateAnalyze}
\title{Update analyze}
\usage{
updateAnalyze(result, pitch_true, spectrogram = NULL)
}
\arguments{
\item{result}{the matrix of results returned by analyze()}

\item{pitch_true}{manual pitch contour of length nrow(result), with NAs}

\item{spectrogram}{spectrogram with ncol = nrow(result)}
}
\description{
Internal soundgen function
}
\details{
Updates the output of analyze using manual pitch. Called by pitch_app().
}
\keyword{internal}
