% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{validatePars}
\alias{validatePars}
\title{Validate parameters}
\usage{
validatePars(p, gp, def, invalidArgAction = c("adjust", "abort", "ignore")[1])
}
\arguments{
\item{p}{parameter name}

\item{gp}{parameter value}

\item{def}{matrix or dataframe containing reference values (low, high,
default)}

\item{invalidArgAction}{what to do if an argument is invalid or outside the
range: 'adjust' = reset to default value, 'abort' = stop execution,
'ignore' = throw a warning and continue (may crash)}
}
\description{
Internal soundgen function
}
\details{
Checks whether the value of a numeric parameter falls within the allowed
range. Options: abort, reset to default, throw a warning and continue.
}
\keyword{internal}
