% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftFormants.R
\name{warpMatrix}
\alias{warpMatrix}
\title{Warp matrix}
\usage{
warpMatrix(m, scaleFactor, interpol = c("approx", "spline")[1])
}
\arguments{
\item{m}{matrix (rows = frequency bins, columns = time)}

\item{scaleFactor}{1 = no change, >1 = raise formants}

\item{interpol}{interpolation method}
}
\description{
Internal soundgen function
}
\details{
Warps or scales each column of a matrix (normally a spectrogram).
}
\examples{
a = matrix(1:12, nrow = 4)
a
soundgen:::warpMatrix(a, 1.5, 'approx')
soundgen:::warpMatrix(a, 1/1.5, 'spline')
}
\keyword{internal}
