% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeAudio}
\alias{writeAudio}
\title{Write audio}
\usage{
writeAudio(x, audio, filename)
}
\arguments{
\item{x}{numeric vector}

\item{audio}{list returned by \code{\link{readAudio}} containing
samplingRate, bit, scale, scale_used}

\item{filename}{full path and filename including .wav}
}
\description{
Internal soundgen function.
}
\details{
Writes a .wav file to disk based on the bit/scale/samplingRate contained in
the internally generated \code{audio} object. The point with using this
function is to package tuneR::Wave + normalize + writeWave in a
soundgen-specific way. Unlike seewave::savewav, writeAudio does NOT normalize
or rescale the input.
}
\keyword{internal}
