% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{zeroOne}
\alias{zeroOne}
\title{Normalize 0 to 1}
\usage{
zeroOne(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{na.rm}{if TRUE, removed NA's when calculating min/max for normalization}
}
\description{
Internal soundgen function
}
\details{
Normalized input vector to range from 0 to 1
}
\keyword{internal}
