% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{annotation_app}
\alias{annotation_app}
\title{Annotation app}
\usage{
annotation_app()
}
\description{
Starts a shiny app for annotating audio. This is a simplified version of
\code{\link{formant_app}} that also works on Macs.
}
\examples{
\dontrun{
# Set up Firefox as default browser either system-wide or just in R.
# For ex., on Linux, run:
options('browser' = '/usr/bin/firefox')  # path to the executable
annotation_app()  # runs in Firefox
}
}
