% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{annotation_app}
\alias{annotation_app}
\title{Annotation app}
\usage{
annotation_app()
}
\description{
Starts a shiny app for annotating audio. This is a simplified and faster
version of \code{\link{formant_app}} intended only for making annotations.
}
\examples{
\dontrun{
annotation_app()  # runs in default browser such as Firefox or Chrome

# To change system default browser, run something like:
options('browser' = '/usr/bin/firefox')  # path to the executable on Linux
}
}
