% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{convert_sec_to_hms}
\alias{convert_sec_to_hms}
\title{Print time}
\usage{
convert_sec_to_hms(time_s, digits = 0)
}
\arguments{
\item{time_s}{time (s)}

\item{digits}{number of digits to preserve for s (1-60 s)}
}
\value{
Returns a character string like "1 h 20 min 3 s"
}
\description{
Internal soundgen function.
}
\details{
Converts time in seconds to time in y m d h min s for pretty printing.
}
\examples{
time_s = c(.0001, .01, .33, .8, 2.135, 5.4, 12, 250, 3721, 10000,
           150000, 365 * 24 * 3600 + 35 * 24 * 3600 + 3721)
soundgen:::convert_sec_to_hms(time_s)
soundgen:::convert_sec_to_hms(time_s, 2)
}
\keyword{internal}
