% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{convert_sec_to_hms}
\alias{convert_sec_to_hms}
\title{Print time}
\usage{
convert_sec_to_hms(time_s)
}
\arguments{
\item{time_s}{time (s)}
}
\value{
Returns a character string like "1 h 20 min 3 s"
}
\description{
Internal soundgen function.
}
\details{
Converts time in seconds to time in hh:mm:ss for pretty printing.
}
\examples{
time_s = c(0, .65, 180, 182, 3721, 10000, 150000,
           365 * 24 * 3600,
           365 * 24 * 3600 + 35 * 24 * 3600 + 3721)
soundgen:::convert_sec_to_hms(time_s)
}
\keyword{internal}
