% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_segment.R
\name{findSyllables}
\alias{findSyllables}
\title{Find syllables}
\usage{
findSyllables(envelope, timestep, threshold, shortestSyl, shortestPause,
  mergeSyl)
}
\arguments{
\item{envelope}{downsampled amplitude envelope}

\item{timestep}{time difference between two points in the envelope (ms)}

\item{threshold}{all continuous segments above this value are considered to
be syllables}

\item{shortestSyl}{minimum acceptable length of syllables, ms}

\item{shortestPause}{minimum acceptable break between syllables, ms.
Syllables separated by less time are merged. To avoid merging, specify
\code{shortestPause = NA}}

\item{mergeSyl}{if TRUE, syllable separated by less than
\code{shortestPause} will be merged}
}
\value{
Returns a dataframe with timing of syllables.
}
\description{
Internal soundgen function.
}
\details{
Called by \code{\link{segment}}.
}
\keyword{internal}
