% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundgenShiny.R
\name{formant_app}
\alias{formant_app}
\title{Interactive formant tracker}
\usage{
formant_app()
}
\description{
Starts a shiny app for manually correcting formant measurements. Supported browsers: Firefox / Chrome. Note that the
browser has to be able to playback WAV audio files, otherwise there will be
no sound.
}
