% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_soundgen.R
\name{htmlPlots}
\alias{htmlPlots}
\title{HTML for clickable plots}
\usage{
htmlPlots(htmlFile, plotFiles, audioFiles = "", width = "900px")
}
\arguments{
\item{width}{default flex-basis, ie approximate default width of each image}

\item{myfolder}{full path to target folder, without a '/' at the end}

\item{myfiles}{a list of full names of files (with paths and extensions)}
}
\description{
Internal soundgen function
}
\details{
Writes an html file for displaying clickable plots in a browser.
}
\examples{
\dontrun{
htmlPlots(myfolder = '~/Downloads/temp',
          myfiles = c('~/Downloads/temp/myfile1.wav',
                      '~/Downloads/temp/myfile2.wav'))
}
}
\keyword{internal}
