% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{hz2mel}
\alias{hz2mel}
\title{Hz to mel}
\usage{
hz2mel(f, htk = FALSE)
}
\arguments{
\item{f}{frequency, Hz}

\item{htk}{algrithm}
}
\description{
Internal soundgen function: a temporary fix needed because tuneR::hz2mel
doesn't accept NAs or vectors.
}
\examples{
soundgen:::hz2mel(c(440, 220, NA))
}
\keyword{internal}
