% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{intplNA}
\alias{intplNA}
\title{Interpolate NAs}
\usage{
intplNA(x, idx_na = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{idx_na}{which(is.na(x))}
}
\value{
Returns the same numeric vector with NAs filled in by interpolation.
}
\description{
Internal soundgen function
}
\details{
Takes a numeric vector and fills in the NAs by linear
interpolation in the middle and constant interpolation at the ends.
}
\examples{
soundgen:::intplNA(c(NA, 405, 441, 460, NA, NA, NA, 480, 490, NA, NA))
}
\keyword{internal}
