% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulationSpectrum.R
\name{modulationSpectrum}
\alias{modulationSpectrum}
\title{Modulation spectrum}
\usage{
modulationSpectrum(x, samplingRate = NULL, dynamicRange = 80,
  windowLength = 25, step = NULL, overlap = 75, wn = "gaussian",
  zp = 0, power = FALSE, plot = TRUE, quantiles = c(0.8, 0.9,
  0.99), kernelSize = 5, kernelSD = 0.5)
}
\arguments{
\item{x}{folder, path to a wav/mp3 file, a numeric vector representing a
waveform, or a list of numeric vectors}

\item{samplingRate}{sampling rate of \code{x} (only needed if
\code{x} is a numeric vector, rather than an audio file)}

\item{dynamicRange}{dynamic range, dB. All values more than one dynamicRange
under maximum are treated as zero}

\item{windowLength}{length of FFT window, ms}

\item{step}{you can override \code{overlap} by specifying FFT step, ms}

\item{overlap}{overlap between successive FFT frames, \%}

\item{wn}{window type: gaussian, hanning, hamming, bartlett, rectangular,
blackman, flattop}

\item{zp}{window length after zero padding, points}

\item{power}{if TRUE, returns power modulation spectrum (^2)}

\item{plot}{if TRUE, plots the modulation spectrum}

\item{quantiles}{quantiles labeled as separate contours on the plot}

\item{kernelSize}{the size of Gaussian kernel used for smoothing}

\item{kernelSD}{the SD of Gaussian kernel used for smoothing, relative to its
size}
}
\description{
Produces a modulation spectrum. Bla-bla-bla...
}
\examples{
ss = vector('list', 10)
for (i in 1:length(ss)) {
  ss[[i]] = soundgen(sylLen = runif(1, 100, 1000), temperature = .4, pitch = runif(3, 400, 600))
}
# lapply(ss, playme)
ms = modulationSpectrum(ss[[1]], samplingRate = 16000)
ms = modulationSpectrum(ss, samplingRate = 16000)

ms = modulationSpectrum('~/Downloads/temp/200_ut_fear-bungee_11.wav')
ms = modulationSpectrum('/home/allgoodguys/Downloads/temp', digits = 5)

ms = modulationSpectrum(ss, samplingRate = 16000, quantiles = c(.8, .9, .99), kernelSize = 6, power = FALSE)
}
\references{
\itemize{
  \item Singh, N. C., & Theunissen, F. E. (2003). Modulation spectra of
  natural sounds and ethological theories of auditory processing. The Journal
  of the Acoustical Society of America, 114(6), 3394-3411.
  }
}
