% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertModulSpec.R
\name{msToSpec}
\alias{msToSpec}
\title{Modulation spectrum to spectrogram}
\usage{
msToSpec(ms, samplingRate, windowLength = NULL, step = NULL)
}
\arguments{
\item{ms}{target modulation spectrum (matrix of complex numbers)}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector, rather than an audio file)}

\item{windowLength}{length of FFT window, ms}

\item{step}{you can override \code{overlap} by specifying FFT step, ms}
}
\value{
Returns a spectrogram - a numeric matrix of complex numbers of
  the same dimensions as ms.
}
\description{
Internal soundgen function.
}
\details{
Takes a complex MS and transforms it to a complex spectrogram with proper row
(frequency) and column (time) labels.
}
\keyword{internal}
