% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notesDict}
\alias{notesDict}
\title{Conversion table from Hz to semitones above C0 to musical notation}
\format{An object of class \code{data.frame} with 192 rows and 2 columns.}
\usage{
notesDict
}
\description{
A dataframe of 132 rows and 2 columns: "note" and "freq" (Hz)
}
\keyword{datasets}
