% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_loudness.R
\name{phon2sone}
\alias{phon2sone}
\title{Convert phon to sone}
\usage{
phon2sone(phon)
}
\arguments{
\item{phon}{loudness level, phon (vectorized)}
}
\description{
Internal soundgen function
}
\details{
Source: Timoney, J., Lysaght, T., Schoenwiesner, M., & MacManus, L. (2004).
Implementing loudness models in matlab.
}
\examples{
phon = 0:120
sone = soundgen:::phon2sone(phon)
plot(phon, sone, type = 'b')
plot(phon, log2(sone), type = 'b')
}
\keyword{internal}
