% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_pitch_postprocessing.R
\name{pitchDescriptives}
\alias{pitchDescriptives}
\title{Pitch descriptives}
\usage{
pitchDescriptives(
  x,
  step = NULL,
  smoothBW = c(NA, 25, 2),
  inflThres = 0.25,
  timeUnit = c("s", "ms")[1],
  plot = FALSE
)
}
\arguments{
\item{x}{input: numeric vector, a list of time stamps and values in rows, a
dataframe with one row per file and time/pitch values stored as characters
(as exported by \code{\link{pitch_app}}), or path to csv file containing
the output of \code{\link{pitch_app}} or \code{\link{analyze}}}

\item{step}{distance between values in s (only needed if input is a vector)}

\item{smoothBW}{a vector of bandwidths (Hz) for consecutive smoothing of
input using \code{\link{pitchSmoothPraat}}}

\item{inflThres}{minimum difference (in semitones) between consecutive
extrema to consider them inflections; to apply a different threshold at
each smoothing level, provide \code{inflThres} as a vector of the same
length as \code{smoothBW}}

\item{timeUnit}{specify whether the time stamps (if any) are in ms or s}

\item{plot}{if TRUE, plots the inflections for manual verification}
}
\value{
Returns a dataframe with columns containing summaries of one or
  multiple inputs (one input per row). The descriptives are as follows:
  \describe{\item{duration}{total duration, s} \item{durDefined}{duration
  after omitting leading and trailing NAs} \item{propDefined}{proportion of
  input with non-NA value, eg proportion of voiced frames if the input is
  pitch} \item{start, end}{the first and last values} \item{mean, median,
  max, min}{average and extreme values} \item{time_max, time_min}{the
  location of minimum and maximum relative to durDefined, 0 to 1}
  \item{range, range_sem, sd, sd_sem}{range and standard deviation on the
  original scale and in semitones} \item{CV}{coefficient of variation =
  mean/sd (provided for historical reasons)} \item{meanSlope,
  meanSlope_sem}{mean slope in Hz/s or semitones/s (NB: does not depend on
  duration or missing values)} \item{meanAbsSlope, meanAbsSlope_sem}{mean
  absolute slope (modulus, ie rising and falling sections no longer cancel
  out)} \item{maxAbsSlope, maxAbsSlope_sem}{the steepest slope}}
}
\description{
Provides common descriptives of time series such as pitch contours, including
measures of average / range / variability / slope / inflections etc. Several
degrees of smoothing can be applied consecutively.
}
\examples{
x = 5 + sin(2 * pi * (1:100) / 15) * seq(1, 5, length.out = 100)
idx_na = c(1:4, 6, 7, 14, 25, 30:36, 39, 40, 42, 45:50,
           57, 59, 62, 66, 71:79, 98)
x[idx_na] = NA
plot(x, type = 'b')
pd = pitchDescriptives(x, step = .025, smoothBW = c(NA, 25, 2),
                       inflThres = c(NA, 2, 2), plot = TRUE)
pd

\dontrun{
# a single file
data(sheep, package = 'seewave')
a = analyze(sheep)
pd1 = pitchDescriptives(a$detailed[, c('time', 'pitch')],
                        timeUnit = 'ms', inflThres = NULL, plot = TRUE)
pd2 = pitchDescriptives(a$detailed[, c('time', 'pitch')],
                        timeUnit = 'ms', inflThres = c(0.1, .5), plot = TRUE)

# multiple files returned by pitch_app()
pd = pitchDescriptives(
  '~/Downloads/pitch_manual_1708.csv',
  timeUnit = 'ms', smoothBW = c(NA, 2), inflThres = .25)

# a single file, exported from Praat
pd = pitchDescriptives(
  '~/Downloads/F-Hin-Om_jana.wav_F0contour.txt',
  timeUnit = 's', smoothBW = c(NA, 2), inflThres = .25, plot = TRUE)
}
}
