% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{processAudio}
\alias{processAudio}
\title{Process audio}
\usage{
processAudio(
  x,
  samplingRate = NULL,
  scale = NULL,
  from = NULL,
  to = NULL,
  funToCall,
  myPars = list(),
  var_noSummary = NULL,
  reportEvery = NULL,
  savePlots = NULL,
  saveAudio = NULL
)
}
\arguments{
\item{x}{path to a folder, one or more wav or mp3 files c('file1.wav',
'file2.mp3'), Wave object, numeric vector, or a list of Wave objects or
numeric vectors}

\item{samplingRate}{sampling rate of \code{x} (only needed if \code{x} is a
numeric vector)}

\item{scale}{maximum possible amplitude of input used for normalization of
input vector (only needed if \code{x} is a numeric vector)}

\item{from}{if NULL (default), analyzes the whole sound, otherwise
from...to (s)}

\item{to}{if NULL (default), analyzes the whole sound, otherwise
from...to (s)}

\item{funToCall}{function to call (specify what to do with each audio input)}

\item{myPars}{a list of parameters to pass on to `funToCall`}

\item{var_noSummary}{names of output variables that should not be summarized}

\item{reportEvery}{report estimated time left every ... iterations (NA = no
reporting, NULL = default frequency)}

\item{savePlots}{full path to the folder in which to save the plots (NULL =
don't save, '' = same folder as audio)}

\item{summaryFun}{function(s) used to summarize the output per input}
}
\description{
Internal soundgen function.
}
\keyword{internal}
