% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pseudoLog_undo}
\alias{pseudoLog_undo}
\title{Undo pseudolog}
\usage{
pseudoLog_undo(x, sigma, base)
}
\arguments{
\item{x}{numeric vector to transform}

\item{sigma}{scaling factor for the linear part}

\item{approximate}{logarithm base used}
}
\description{
Internal soundgen function
}
\details{
From scales::pseudo_log_trans()
}
\seealso{
\code{\link{pseudoLog}}
}
\keyword{internal}
