% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{semitonesToHz}
\alias{semitonesToHz}
\title{Convert semitones to Hz}
\usage{
semitonesToHz(s, ref = 0.5109875)
}
\arguments{
\item{s}{vector or matrix of frequencies (semitones above C0)}

\item{ref}{frequency of the reference value (defaults to C-5, 0.51 Hz)}
}
\description{
Converts from semitones above C-5 (~0.5109875 Hz) to Hz. See
\code{\link{HzToSemitones}}
}
