% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{spectrogramFolder}
\alias{spectrogramFolder}
\title{Spectrogram folder}
\usage{
spectrogramFolder(...)
}
\arguments{
\item{...}{any input parameters}
}
\description{
Deprecated; use \code{\link{spectrogram}} instead
}
