% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralDescr.R
\name{subhToHarm}
\alias{subhToHarm}
\title{Subharmonics-to-harmonics ratio}
\usage{
subhToHarm(
  frame,
  bin,
  freqs,
  pitch,
  samplingRate,
  tol = 0.05,
  nHarm = 5,
  harmThres = 3,
  harmTol = 0.25,
  method = c("cep", "harm")[1]
)
}
\description{
Internal soundgen function
}
\keyword{internal}
