% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_math.R
\name{switchColorTheme}
\alias{switchColorTheme}
\title{Switch color theme}
\usage{
switchColorTheme(colorTheme)
}
\arguments{
\item{colorTheme}{string like 'bw', 'seewave', or function name}
}
\description{
Internal soundgen function
}
\keyword{internal}
