% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analyze.R
\name{updateAnalyze}
\alias{updateAnalyze}
\title{Update analyze}
\usage{
updateAnalyze(
  result,
  pitch_true,
  pitchCands_list = NULL,
  spectrogram,
  freqs = NULL,
  bin = NULL,
  samplingRate = NULL,
  harmHeight_pars = list(),
  subh_pars = list(),
  smooth,
  smoothing_ww,
  smoothingThres,
  varsToUnv = NULL
)
}
\arguments{
\item{result}{the matrix of results returned by analyze()}

\item{pitch_true}{manual pitch contour of length nrow(result), with NAs}

\item{spectrogram}{spectrogram with ncol = nrow(result)}

\item{freqs}{frequency labels of spectrogram bins}

\item{bin}{spectrogram bin width}

\item{harmHeight_pars}{same as argument "harmHeight" to analyze() - a list of
settings passed to soundgen:::harmHeight()}

\item{smooth, smoothing_ww, smoothingThres}{smoothing parameters}

\item{varsToUnv}{set these variables to NA in unvoiced frames}
}
\description{
Internal soundgen function
}
\details{
Updates the output of analyze using manual pitch. Called by pitch_app().
}
\keyword{internal}
